Add desktop files
Currently, generate these ourselves based on a template.
Not using the GOG-provided ones as these have proven difficult to
get since we need to run the installer non-interactively, which
appears to require specifying "--nooptions", which prevents
desktop files from being installed.

Perhaps in the future a better way will be found.