Set executable bit on game folder
Previously GOG shipped games with all files having the exec bit set.
Not unreasonably they don't chmod +x everything any longer, but
they also don't set x on the binaries that actually need it, instead
doing that as part of the start.sh script. Since the contents of the
flatpak are read-only, that script fails.

As a not-perfect workaround, we chmod +x the entire game dir so at
least things can start again.